move memory management to personal settings page
