+++ packages/frontend/@n8n/i18n/src/locales/en.json
@@ -2722,7 +2722,7 @@
 	"settings.chatHub.providers.table.action.editProvider": "Edit provider",
 	"settings.chatHub.providers.table.action.editProvider.disabled": "Only instance admins and owners can edit chat providers.",
 	"settings.chatHub.memory.title": "Memory",
-	"settings.chatHub.memory.description": "Facts remembered from past conversations",
+	"settings.chatHub.memory.description": "Facts remembered from past conversations in ChatHub",
 	"settings.chatHub.memory.empty": "No facts remembered yet",
 	"settings.chatHub.memory.delete.tooltip": "Forget this",
 	"settings.chatHub.memory.delete.error": "Error deleting memory",
@@ -4374,6 +4374,7 @@
 	"settings.personal.theme.systemDefault": "System default",
 	"settings.personal.theme.light": "Light theme",
 	"settings.personal.theme.dark": "Dark theme",
+	"settings.personal.chatHubMemory": "ChatHub Memory",
 	"settings.personal.mfa.button.disabled.infobox": "Two-factor authentication is currently disabled.",
 	"settings.personal.mfa.button.enabled.infobox": "Two-factor authentication is currently enabled.",
 	"settings.personal.mfa.button.enabled": "Enable 2FA",
